tests/wan2_2: drop non-production wav2vec2-base test, rename large→encoder
Of the two wav2vec2 PCC tests, only ``test_wav2vec2_large_encoder.py``
exercises the production audio encoder (``wav2vec2-large-xlsr-53-english``
bundled in ``Wan-AI/Wan2.2-S2V-14B``, with ``feat_extract_norm="layer"``
+ ``do_stable_layer_norm=True``). ``test_wav2vec2_encoder.py`` was
testing the base ``facebook/wav2vec2-base-960h`` variant which the S2V
pipeline never loads — drop it.

The production test is renamed from ``test_wav2vec2_large_encoder.py``
to ``test_wav2vec2_encoder.py`` (the ``_large`` suffix was only needed
to distinguish it from the now-deleted base variant). Function renamed
to ``test_wav2vec2_encoder`` to match.
