+++ models/tt_dit/tests/models/wan2_2/test_wav2vec2_encoder.py
@@ -2,6 +2,13 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+"""PCC test for wav2vec2-large-xlsr-53-english using the weights bundled in
+``Wan-AI/Wan2.2-S2V-14B`` (the production audio encoder). Exercises the
+``feat_extract_norm="layer"`` + ``do_stable_layer_norm=True`` code paths.
+"""
+
+from pathlib import Path
+
 import pytest
 import torch
 from loguru import logger
@@ -18,11 +25,28 @@ from ....utils.tensor import to_torch
 from ....utils.test import line_params
 
 
+# Resolve the bundled large-xlsr-53 weights inside the S2V download.
+def _bundled_wav2vec2_path() -> Path | None:
+    base = Path.home() / ".cache/huggingface/hub/models--Wan-AI--Wan2.2-S2V-14B/snapshots"
+    if not base.exists():
+        return None
+    snaps = list(base.iterdir())
+    if not snaps:
+        return None
+    candidate = snaps[0] / "wav2vec2-large-xlsr-53-english"
+    return candidate if candidate.exists() else None
+
+
+_BUNDLED_PATH = _bundled_wav2vec2_path()
+
+
+@pytest.mark.skipif(
+    _BUNDLED_PATH is None,
+    reason="Wan-AI/Wan2.2-S2V-14B bundled wav2vec2 weights not found. Run `hf download Wan-AI/Wan2.2-S2V-14B`.",
+)
 @pytest.mark.parametrize(
     ("mesh_device", "tp_axis", "num_links", "device_params", "topology"),
     [
-        pytest.param((2, 2), 1, 1, line_params, ttnn.Topology.Linear, id="2x2_tp1"),
-        pytest.param((2, 4), 1, 1, line_params, ttnn.Topology.Linear, id="2x4_tp1"),
         pytest.param((4, 8), 0, 2, line_params, ttnn.Topology.Linear, id="bh_4x8_tp0"),
     ],
     indirect=["mesh_device", "device_params"],
@@ -30,10 +54,7 @@ from ....utils.test import line_params
 @pytest.mark.parametrize(
     "audio_samples",
     [
-        # ~1.024s at 16 kHz. Short clip keeps the test fast while exercising the
-        # full feature extractor + all 12 encoder layers.
         pytest.param(16000, id="1s"),
-        pytest.param(32000, id="2s"),
     ],
 )
 def test_wav2vec2_encoder(
@@ -43,27 +64,23 @@ def test_wav2vec2_encoder(
     topology: ttnn.Topology,
     audio_samples: int,
 ) -> None:
-    model_id = "facebook/wav2vec2-base-960h"
+    model_path = str(_BUNDLED_PATH)
+    logger.info(f"Loading wav2vec2-large-xlsr-53 from bundled S2V path: {model_path}")
 
-    # 1. Load HF reference (CPU, float32 for the golden).
-    hf_model = Wav2Vec2Model.from_pretrained(model_id).eval()
-    processor = Wav2Vec2Processor.from_pretrained(model_id)
+    hf_model = Wav2Vec2Model.from_pretrained(model_path).eval()
+    processor = Wav2Vec2Processor.from_pretrained(model_path)
     config = Wav2Vec2Config.from_hf(hf_model.config)
+    assert config.do_stable_layer_norm, "Expected do_stable_layer_norm=True for large-xlsr-53"
+    assert config.feat_extract_norm == "layer", "Expected feat_extract_norm='layer' for large-xlsr-53"
 
-    # 2. Build a deterministic synthetic waveform and normalize via the HF
-    # processor so the TT path sees the same input distribution as the
-    # production pipeline.
     torch.manual_seed(0)
     waveform = torch.randn(audio_samples)
-    inputs = processor(waveform.numpy(), sampling_rate=16000, return_tensors="pt")
-    input_values = inputs.input_values  # [1, T_raw]
+    input_values = processor(waveform.numpy(), sampling_rate=16000, return_tensors="pt").input_values
 
-    # 3. Golden forward.
     with torch.no_grad():
-        golden = hf_model(input_values).last_hidden_state  # [1, T_features, 768]
+        golden = hf_model(input_values).last_hidden_state
     logger.info(f"Golden hidden shape: {tuple(golden.shape)}")
 
-    # 4. Build TT model.
     tp_factor = tuple(mesh_device.shape)[tp_axis]
     parallel_config = EncoderParallelConfig(
         tensor_parallel=ParallelFactor(mesh_axis=tp_axis, factor=tp_factor),
@@ -79,11 +96,7 @@ def test_wav2vec2_encoder(
     tt_model.load_torch_state_dict(hf_model.state_dict())
     tt_model.bind_cpu_modules(hf_model)
 
-    # 5. Forward — feature extractor + feature_projection + 12 transformer
-    # layers on device; pos-conv + initial LayerNorm on CPU
-    # (see Wav2Vec2Encoder docstring for the kernel constraint that requires this).
     tt_hidden = tt_model(input_values)
-    tt_hidden_torch = to_torch(tt_hidden)  # replicated → just take the first shard
-
+    tt_hidden_torch = to_torch(tt_hidden)
     logger.info(f"TT hidden shape: {tuple(tt_hidden_torch.shape)}")
-    assert_quality(tt_hidden_torch, golden, pcc=0.99)
+    assert_quality(tt_hidden_torch.float(), golden, pcc=0.99)
